V3: Angular 8 / Java 11
* Drop AngularJS and Thymleaf
* Use Clarity Design instead of Boostrap
* Use Spring-boot 2.2 instead of 1.5

+++ trading-common/src/main/java/fr/ymanvieu/trading/util/MathUtils.java
@@ -38,8 +38,8 @@ public class MathUtils {
 		return divide(BigDecimal.ONE, d);
 	}
 
-	public static float percentChange(BigDecimal oldValue, BigDecimal newValue) {
-		return divide(newValue.subtract(oldValue), oldValue).multiply(HUNDRED).floatValue();
+	public static double percentChange(BigDecimal oldValue, BigDecimal newValue) {
+		return divide(newValue.subtract(oldValue), oldValue).multiply(HUNDRED).doubleValue();
 	}
 
 	public static boolean equalsByComparingTo(BigDecimal bd1, BigDecimal bd2) {
